Rename and split programming error tests
