+++ tests/archivey/test_usage_errors.py
@@ -1,7 +1,3 @@
-import struct
-import zipfile
-from datetime import datetime, timezone
-
 import pytest
 
 from archivey.core import open_archive
@@ -66,13 +62,3 @@ def test_resolve_link_regular_file(sample_archive: SampleArchive, sample_archive
         assert archive.resolve_link(member) is member
 
 
-def test_zip_extra_field_before_timestamp(tmp_path) -> None:
-    path = tmp_path / "extra.zip"
-    modtime = int(datetime(2020, 1, 2, 3, 4, 5, tzinfo=timezone.utc).timestamp())
-    zi = zipfile.ZipInfo("file.txt", date_time=(2020, 1, 2, 3, 4, 5))
-    zi.extra = struct.pack("<HH4s", 0x1234, 4, b"abcd") + struct.pack("<HHB", 0x5455, 5, 1) + struct.pack("<I", modtime)
-    with zipfile.ZipFile(path, "w") as zf:
-        zf.writestr(zi, b"data")
-    with open_archive(str(path)) as archive:
-        info = archive.get_members()[0]
-        assert info.mtime_with_tz == datetime(2020, 1, 2, 3, 4, 5, tzinfo=timezone.utc)
